requirements.txt
tensorflow==2.2.0
numpy==1.18.5
Pillow==8.1.0
